admin endpoints to populate tables for dynamic k8s clusters
Adds 'admin/lando-connections' that allows creating and reading LandoConnections
Adds 'admin/job-settings' that allows creating and reading JobSettings
Adds 'admin/job-strategies' that allows creating, deleting and reading JobStrategies.
The LandoConnection and JobSettings are directly referenced by Jobs so
cannot be safely deleted. JobStrategy is not tied directly to a Job
so can be deleted.
NOTE: If JobStrategy is used as a default_job_strategy in a WorkflowConfiguration
then job strategy delete will fail. The intension with this change is that users
of dynamic clusters will specify the job strategy created for the dynamic cluster.